Rename field for filtering

+++ evelon-demo/src/main/java/dev/httpmarco/evelon/demo/models/AbstractModel.java
@@ -12,7 +12,7 @@ public final class AbstractModel extends TestObject1 {
 
     private final UUID test;
 
-    public AbstractModel(String permission, long expire, UUID test) {
+    public AbstractModel(String permission, int expire, UUID test) {
         super(permission, expire);
         this.test = test;
     }
